requirements.txt
rdflib
owlready2
requests
openai
jupyter
ipykernel
pyshacl
python-dotenv